Hooks refactoring
* more order in constants
* snackbars may have function to get message text
