--- src/components/compare-button/index.js
@@ -1,1 +1,0 @@
-export {CompareButton} from './compare-button'
