+++ src/components/index.js
@@ -2,4 +2,4 @@ export {PriceItem} from './price-item';
 export {Input} from './input';
 export {MeasuresList} from './measures-list';
 export {Header} from './header';
-export {CompareButton} from './compare-button';
+export {TopButtonsBlock} from './top-buttons-block';
