+++ src/components/top-buttons-block/index.js
@@ -1,0 +1,1 @@
+export {TopButtonsBlock} from './top-buttons-block'
