+++ src/constants/field-names.js
@@ -1,11 +1,6 @@
-export const COMPARE_DATA = 'compareData';
 export const UNIT = 'unit';
 export const PRICE = 'price';
 export const QUANTITY = 'quantity';
-export const BEST_VALUES_INDEXES = 'bestValuesIndexes';
-export const MEASURE = 'measure';
-export const SIDEBAR_VISIBLE = 'sidebarVisible';
-export const DISPLAY_SNACKBARS = 'displaySnackBars';
 
 export const SNACKBAR_MINIMUM_TWO = 'snackbarMinimumTwo';
 export const SNACKBAR_NOT_FILLED_DATA = 'snackbarNotFilledData';
@@ -13,11 +8,13 @@ export const SNACKBAR_NOT_FILLED_DATA = 'snackbarNotFilledData';
 export const SNACKBARS_LIST = [
     {
         key: SNACKBAR_MINIMUM_TWO,
-        text: 'Для сравнения необходимо не менее 2 позиций'
+        text: 'Для сравнения необходимо не менее 2 позиций',
     },
     {
         key: SNACKBAR_NOT_FILLED_DATA,
-        text: 'Не все поля заполнены полностью'
+        getErrorText: (errors) => {
+            return `${errors.length} полей не заполнено`
+        }
     }
 ];
